third_party/WebKit/LayoutTests/css-parser/font-variant-east-asian-parsing.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/property-parsing-test.js"></script>
<script>
// Verifies that font-variant-east-asian property and its value are properly parsed

assert_valid_value("font-variant-east-asian", "ruby full-width jis78", "jis78 full-width ruby");
</script>
